Fixed setup file and restructured plugins in the course.
The setup file can now install p1tr globally, so you can launch it with the
p1tr command. A config file must be present in the working directory, or the
path to one must be provided with the -c command line switch.

In order to have P1tr usable both in installed and standalone mode, some
modules were moved around and imports changed. To make installed plugins
available to users of the p1tr command, support for loading plugins from
the install directory was added.

Priority treatment for plugins in the current working directory's plugin folder
was made possible by changing sys.path at the beginning of p1tr.py to use the
current working directory as the primary search path.