Add wrapper for Python's logging module
The purpose of this wrapper is to simplify writing to logs for plugins and the bot core
itself. Its interface is similar to the essentials of the logging module. For a plugin
developer, the functions debug, info, warning, error and critical are relevant. They
take a message as the first argument. Keyword arguments can follow. If none are specified,
the message is sent to the global log. If a server is specified, the server log is filled.
If server and channel are specified, the message is sent to the channel's log. The name
of a plugin can also be supplied to write a plugin log.

Although the logging module provides ways to set separate loglevels for each handler, the
file- and stream handler used in this wrapper take whatever goes through the logger's
filter.