Create base class for authorization providers
Authorization providers are special plugins, whose methods are called
when privileged commands are issued by a user. There are the following
ranks:
[normal] < Voice < Half-OP < OP < Channel Owner < Master
A higher rank can do whatever the ranks below it can do as well.

A first authorization provider is in the works, based on freenode's
NickServ and ChanServ services. One more provider is planned, which
should be network-independent.

Customization of the commands' authorization requirements (so that
the annotations are just defaults) is not in the works yet. A good
location for that is likely the authorization provider base class.