Adapt authorization provider base class to permit deferred command execution
If authorization requires the provider to launch an asynchronous request to,
let's say, NickServ, the authorization outcome can't be provided via return
value.

To counter this problem, the affected plugin, name of the command and parameters
are supplied to the authorization methods. It becomes the authorization provider's
responsibility to execute privileged commands as soon as authorization is
completed successfully, or provide an error message otherwise.

Normal commands, which don't require special privileges, are not affected.