Add support for require_authenticated
Beware: Right now, inferior ranks are not included in superior ones. E.g. if you're
master and try to execute an op command, you can't do that. The responsibility to
provide this behavior has been delegated to auth providers, since various approaches
have to deal with this issue very differently.