Progress on the NickServ-/ChanServ-based authorization provider
The ranks Master, OP, Half-OP, Voice, and Authenticated are now fully usable.
The missing feature that superior ranks can also execute commands which require
a lower rank was also added for the ranks mentioned above.

Right now, it is necessary, that the user is currently OP etc. in the channels in
question. Accessing the ACCESS LIST of ChanServ is not implemented yet.