Implement plugin unit testing
When running P1tr with the -t option, you can now run unit tests stored in the
plugins.PLUGIN.PLUGIN_test module. Tests are automatically discovered and run,
as long as they are derived from p1tr.test.PluginTestCase, and the methods to
test are denoted with the @test annotation.

At the end of the test run, a summary of the result is provided, along with a
nice green (or red, depending on the outcome) bar. As soon as the tests have
finished, the bot terminates.

From this point on, it is highly encouraged to accompany plugins with a test
suite. The plugin scaffold generation script already generates a sample test
case for you. Existing plugins will be retrofitted with test cases.

Please keep in mind that the testing environment is still very volatile. Additions
affecting how test cases are written may be made in the future, but I'll try not to
break compatibility.