Fix data directory location by introducing initialize method to plugins
Plugin data was previously stored in the current working directory, because the
data_path member of the plugins was set after instantiation of the plugin, but the
storage files were already loaded in the __init__ method - before data_path was set.

As a solution, an initialize method, which is called at plugin load before any other
plugin method (except __init__), is used for plugin initialization now.

Existing plugins have been migrated to the new solution.